Cleaning up code + finishing category module

+++ admin/controllers/users.php
@@ -5,10 +5,14 @@
         $xtpa = new XTemplate("views/users.html");
         $condition = "1 = 1";
         if(isset($_POST['btnDel'])){
-            $ls = implode(', ',$_POST['ck']);
-            $sql = "DELETE FROM users WHERE id IN ({$ls})";
-            $db->execSQL($sql);
+            //If 'Delete All' button is clicked
+            if(isset($_POST['ck'])){
+                $ls = implode(', ',$_POST['ck']);
+                $sql = "DELETE FROM users WHERE id IN ({$ls})";
+                $db->execSQL($sql);
+            }
         }else if(isset($_GET['s'])){
+            //If 'Search Button' is clicked
             $str = $_GET['s'];
             $search_val = str_replace('+','',$str);
             if ($str == 'Male' || $str == 'Female') $condition .= " AND gender = '{$search_val}' ";
@@ -19,6 +23,7 @@
             }
             $xtpa->assign('search',$search_val);
         }
+        //Fetch all entries from DB with condition created above
         $user = $db->fetch("SELECT * FROM users WHERE {$condition}");
         $t = count($user);
         $l = 10;
@@ -26,6 +31,7 @@
         $offset = ($page - 1) * $l;
         $condition .= " ORDER BY id ASC LIMIT {$offset}, {$l} ";
         $url = 'a=users';
+        //Fetch only 10 entries from DB with condition and limit as above
         $user = $db->fetch("SELECT * FROM users WHERE {$condition}");
         $pager = $f->paging($url,$t,$l,'pager', (isset($_GET['s']))?$_GET['s']:'');
         if (count($user) > 0){
